[media] adv7180: Do implicit register paging
The ad7180 has multiple register pages which can be switched between by writing
to a register. Currently the driver manually switches between pages whenever a
register outside of the default register map is accessed and switches back after
it has been accessed. This is a bit tedious and also potential source for bugs.
Having multiple pages means that a register access is not a single atomic
i2c_smbus_write_byte_data() or i2c_smbus_read_byte_data() call and we need to
make sure that concurrent register access does not race against each other. This
patch adds two helper functions that take care of switching between pages and
reading/writing the register. Those functions have a lockdep_assert_held() to
make sure that they are called with the state mutex held to avoid race
conditions. The register numbers for registers are updated to encode both the
page (in the upper 8-bits) and the register (in the lower 8-bits) numbers.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>