[media] adv7180: Do not request the IRQ again during resume
Currently the IRQ is requested from within the init_device() function. This
function is not only called during device probe, but also during resume causing
the driver to try to request the IRQ again. Move requesting the IRQ from
init_device() to the probe function to make sure that it is only requested once.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>